requirements.txt
numpy
pandas
scikit-learn
matplotlib
torch
torchvision
torchaudio
seaborn
tsfresh
tqdm
stumpy
statsmodels
numba
cloudpickle
joblib
jupyter
ipykernel
ipywidgets
networkx
umap-learn